+++ lib/view_component/base.rb
@@ -31,13 +31,13 @@ module ViewComponent
       attr_writer :config
     end
 
+    ViewContextCalledBeforeRenderError = Class.new(StandardError)
+
     include ViewComponent::ContentAreas
     include ViewComponent::PolymorphicSlots
     include ViewComponent::SlotableV2
     include ViewComponent::Translatable
     include ViewComponent::WithContentHelper
-
-    ViewContextCalledBeforeRenderError = Class.new(StandardError)
 
     RESERVED_PARAMETER = :content
 
@@ -230,7 +230,7 @@ module ViewComponent
       if view_context.nil?
         raise(
           ViewContextCalledBeforeRenderError,
-          "`#helpers` can't be used during initialization, as it depends " \
+          "`#helpers` can't be used during initialization as it depends " \
           "on the view context that only exists once a ViewComponent is passed to " \
           "the Rails render pipeline.\n\n" \
           "It's sometimes possible to fix this issue by moving code dependent on " \