+++ test/sandbox/test/rendering_test.rb
@@ -462,6 +462,14 @@ class RenderingTest < ViewComponent::TestCase
     assert_selector("h2", text: I18n.t("translations_component.subtitle"))
   end
 
+  def test_renders_component_with_initializer_translations
+    err =
+      assert_raises ViewComponent::Base::ViewContextCalledBeforeRenderError do
+        render_inline(InitializerTranslationsComponent.new)
+      end
+    assert_includes err.message, "can't be used during initialization"
+  end
+
   def test_renders_component_with_rb_in_its_name
     render_inline(EditorbComponent.new)